refactor(api): PR-5 commit 2 — health/presets/browse/events_sse 4 router 抽出
server.py 4410 → 4068 行（−342），20 routes 搬到 api/routers/ + 4 个 inline
BaseModel 抽到 api/schemas/presets.py。第一批覆盖 4 个域：

  api/routers/health.py        3 routes  /api/health, /api/system/stats, /api/state
  api/routers/presets.py      13 routes  /api/schema + 10 presets CRUD + 2 configs redirect
  api/routers/browse.py        3 routes  /api/datasets, /api/browse, /api/datasets/thumbnail
  api/routers/events_sse.py    1 route   /api/events SSE

  api/schemas/presets.py      4 BaseModel  DuplicateRequest / PresetExportBody /
                                          PresetImportBody / PresetImportFromPathBody
                                          （命名去掉旧 `_` 前缀，统一公开 schema 命名）

api/app.py：4 个 router 一次性 `app.include_router(...)`。

行为零变更：route 路径 / 方法 / 签名 / 返回结构 / 异常码全照抄。Route
snapshot 仍 160（135 server.py @app + 20 include_router + 5 SPA mount/openapi/docs
等非 APIRoute）。

test_route_invariants.py 配合更新：从只数 server.py @app 装饰器改成数 server.py
+ api/routers/*.py 全部 `@router.<verb>` 装饰器之和；PR-6 继续抽 router 时
本测试自动覆盖新文件。

回归：PR-1 安全网 152 + test_studio_server 44 + test_presets_io 20 + 其它
= **232/232 全绿**。route snapshot 不变（key 不变）。
