+++ studio/api/schemas/__init__.py
@@ -1,0 +1,4 @@
+"""Per-router 请求 / 响应 pydantic 模型集中地（PR-5 起从 server.py inline BaseModel 抽出）。
+
+每个文件对应一个 router：api/schemas/presets.py ↔ api/routers/presets.py
+"""
