+++ tests/test_route_invariants.py
@@ -2,10 +2,10 @@
 
 snapshot 是精细网（任何字符变化都触发），本文件是粗粒度二道防线：
 - 数量落在合理区间
-- server.py 源里的 @app.<verb> 装饰器数 == APIRoute 数
+- server.py + api/routers/*.py 的 @app.<verb> / @router.<verb> 装饰器总数 == APIRoute 数
 
-如果某个未来 PR 把 router 拆走了，snapshot 测试会同时挂；但本测试也会挂，
-能在 reviewer 视野里多出一条提醒：APIRoute 数量变了。
+PR-5 起从 server.py 抽 router 到 api/routers/，本测试同步扫描两处装饰器
+之和。新加 router 文件（PR-6 继续抽）后扫描自动覆盖。
 """
 from __future__ import annotations
 
@@ -16,7 +16,9 @@ from fastapi.routing import APIRoute
 
 from studio.server import app
 
-SERVER_PY = Path(__file__).parent.parent / "studio" / "server.py"
+STUDIO_DIR = Path(__file__).parent.parent / "studio"
+SERVER_PY = STUDIO_DIR / "server.py"
+API_ROUTERS_DIR = STUDIO_DIR / "api" / "routers"
 
 
 def test_route_count_in_sane_range() -> None:
@@ -25,18 +27,34 @@ def test_route_count_in_sane_range() -> None:
 
 
 def test_decorator_count_matches_api_routes() -> None:
+    # server.py 里 `@app.<verb>(...)` 装饰器
     src = SERVER_PY.read_text(encoding="utf-8")
-    # 匹配 @app.get(...) / @app.post(...) / ... 以及 @app.api_route(...)
-    decorator_count = len(
+    app_decorator_count = len(
         re.findall(
             r"^@app\.(get|post|put|delete|patch|api_route)\b",
             src,
             flags=re.MULTILINE,
         )
     )
+    # api/routers/*.py 里 `@router.<verb>(...)` 装饰器
+    router_decorator_count = 0
+    if API_ROUTERS_DIR.is_dir():
+        for py in sorted(API_ROUTERS_DIR.glob("*.py")):
+            if py.name == "__init__.py":
+                continue
+            router_src = py.read_text(encoding="utf-8")
+            router_decorator_count += len(
+                re.findall(
+                    r"^@router\.(get|post|put|delete|patch|api_route)\b",
+                    router_src,
+                    flags=re.MULTILINE,
+                )
+            )
+
+    decorator_total = app_decorator_count + router_decorator_count
     api_route_count = sum(1 for r in app.routes if isinstance(r, APIRoute))
-    assert decorator_count == api_route_count, (
-        f"server.py 里 @app.* 装饰器 {decorator_count} 个，"
-        f"但 app.routes 里 APIRoute 实例 {api_route_count} 个 —— "
-        f"差额可能来自 include_router 或 router 注册时丢了一个"
+    assert decorator_total == api_route_count, (
+        f"装饰器总数 {decorator_total}（server.py {app_decorator_count} + "
+        f"api/routers/ {router_decorator_count}）≠ app.routes 里 APIRoute 实例 "
+        f"{api_route_count} —— 差额可能来自漏 include_router 或 router 注册时丢了一个"
     )
